call allocRecord to allocate record type.
This changeset also change the translation of CALL to what it did
before, to generate CALL with less register dependencies.

instruction selection for CALL is broken. Currently it does not list
caller save register as the destination registers. If allocRecord is
replaced with malloc, one test case fails because ecx is trashed by
malloc.

Listing caller-save registers as the destination of CALL requires
spilling for most cases. So the next step is to implement spill.